% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_reveal.R
\name{scroll_reveal}
\alias{scroll_reveal}
\title{Animate elements when scrolling to view them}
\usage{
scroll_reveal(
  target,
  duration = 1000,
  delay = 100,
  distance = "20px",
  reset = T
)
}
\arguments{
\item{target}{The elements to animate as they scroll into view}

\item{duration}{The duration of the animation in milliseconds. Default to 1000 ms}

\item{delay}{The desired delay in milliseconds before triggering the animation. Default to 100 ms}

\item{distance}{Controls how far elements move when revealed. Default to 20px}

\item{reset}{Boolean, should the function animate the element each time it scrolls into view or only once. Defaults to TRUE.}
}
\value{

}
\description{
Animate elements when scrolling to view them
}
