% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_reveal.R
\name{use_reveal}
\alias{use_reveal}
\title{Enable the scrollrevealjs library}
\usage{
use_reveal()
}
\value{
called for the side effect of activating the scrollrevealjs library
}
\description{
The function activates the capabilities of the scrollrevealjs library. The user can put it anywhere within the UI but it's preferable to implement it at the bottom of the UI.
}
\examples{
\donttest{

# Put the function at the bottom of the UI

use_reveal()



}
}
